+++ changelog.md
@@ -6,6 +6,7 @@ FORMAT: 1A
 --------------------
 * ENH: instrumentation added for database, aac, and overall http calls - made visible in /stats for now
 * ENH: If database schema does not match expected, service will startup in readonly mode, and switch to writeable once migration is complete.
+* ENH: Add deadlock configuration parameters to global configurations.
 
 ## Release v1.0.1.26 (June 8, 2017)
 --------------------